Refactor: Rename Medication Utility Files for Clarity
Renames medication utility files to match the functions they export, improving clarity and organization:

- Renamed `schedule-formatting.ts` to `format-schedule.ts`.
- Renamed `next-due-calculation.ts` to `calculate-next-due.ts`.
- Renamed corresponding test files (`format-schedule.test.ts`, `calculate-next-due.test.ts`).
- Updated all import paths in components and test files to reflect these new file names.

All verification steps (formatting, linting, tests, build) passed successfully.
