+++ src/app/medication/utils/format-schedule.ts
